Update how-to-login.md
Updated content to make it more generic to non-training participants. Exercises relevant to training course moved to Exercise 1-1 and 1-2. 

To discuss whether to change the salsadigital.com.au/user example. Do we need authorisation to use an agency's site here? The GovCMS site does allow access to the login page so removed that example.